api/requirements.txt
openai==1.16.2
python-dotenv==1.0.1
fastapi==0.110.1
starlette==0.37.2
click==8.1.7
uvicorn==0.29.0